Added files via upload
Problem 1. (Earthquakes) Consider the earthquakes-2014 dataset (on the cloud). This is taken from a catalog data from the Southern California Earthquake Data Center (SCEDC) holdings for the year 2014. The data include local, regional, and quarry-blast events with epicenters between latitudes 32 degrees S and 37 degrees N and longitudes between -122 degrees W and -114 degrees E. See http://www.data.scec.org for more information. To read the data, use load and the data frame is called dat.

A. Focus on events of magnitude at least 2. Use the function subset to extract the corresponding observations. When this is done, produce a table of counts for the number of events in each month, and then draw a relevant plot.
B. Were earthquakes more prevalent some months out of the year 2014? Formalize this question into a test of hypothesis (what is the null?) and perform a test. Name the test and specify how you choose to calibrate it. After you perform the test, o er some brief comments.

Pr