Update Assignment 2.R
Problem 1.  (A test for symmetry)Consider the setting where we have a numerical sample X1; : : : ; Xn that  is  assumed  iid  from  a  given  distribution.   We  want  to  test  the  distribution  is symmetric  about  0. The  following  is  a  well-known  test  based  on  random  sign  
flipping.   For  a sign vector"= ("1; : : : ; "n)2 f1;+1gn,  compute the sample mean of"1X1; : : : ; "nXn,  denoted Y"=1nPi"iXi.  Let Y?denote the original sample mean.  The one-sided version of the test returns the p-value P= 2n#f":Y"Y?g:(1)
A. Make sense of this in your mind.  To probe your understanding, explain why the factor 2^-n is there.
B. In practice there are too many sign vectors to compute the p-value exactly and one has to resort to Monte Carlo sampling. Write a functionflipSignTest1(x, B=999 )which implements this. x is the vector of observations and Bis the number of MC iterations.
C. How about a two-sided test?  To answer this question, write a functionflipSignTest2(x,B=999 )which  implements the two-sided version.

Problem  2. (Father  and  son's  heights)Consider  the father.son dataset  in  the UsingR package. Suppose we want to compare the father and son's heights. Particularly, say we want to test whether a son tends to be taller than his father.
A. Produce a plot that would help answer this question (at least in an informal way).
B. Denote the data (Xi; Yi); i= 1; : : : ; n, where X is the father's height and Y the son's height .Let Zi = Yi - Xi.  There are several ways to formalize the question into a test of hypotheses.One  of  them  is  to  test  for  symmetry  of  the  distribution  of Z about 0. Use your functionflipSignTest1 to perform such a test.  offer some brief comments.  [This test is in fact a permutation test.  Can you see why?]

Problem 3. (Bootstrap confidence interval) Write code to compute a one-sided 90% boot-strap  pivotal  confidence  interval of the mean difference in  height between son and father. How would that help answer the question in Problem 2? What population are we drawing an inference about?